Moved rift cell data storage to world level
Previously, we were storing all data related to rift cells on the cells themselves. However, that had a number of drawbacks that were limiting the design & interation of that feature. By moving the storage method to the world layer (mimicking how AE2 handles spatial cell data), we get several benefits:
- Much better performance, since data doesn't get serialized to NBT except on save/load
- Clearer separation of concerns between data storage and item class
- Can store output blocks rather than recipes (technically could have done that before, but it's easier to implement now that the code is broken out like this)