Added SpatialRiftBaseRecipe & Manager
Also decided that the pattern I was attempting to use with BidirectionalSerializer was terrible and didn't work, so I abandoned it for this implementation. It turns out that there's a substantial difference between a "real" recipe in memory and a recipe stored on disk, and that difference is that the former must use real registry objects, while the latter can use registry objects that might or might not exist.